egg_info: delay import until needed
When including ansible-lint in a pip requirements.txt file and
installing via `pip -r` the install fails. This is due to pip operating
in two runs. In the first run it executes `python setup.py egg_info` on
each import and then installs the egg in the second run.

Due to the `egg_info` call, ansible-lint had an install-time dependency
on ansible in __init.py__. To avoid this issue, delay importing any
`ansible` or `ansiblelint.utils` modules until when needed.