Ensure matchplay copes with dicts
matchplay is badly named - really it should be matchyaml or something.
It can be run against each play in a play, but it's nice for rules
to be able to just act on the yaml data.

Allow matchplay to still work on yaml that represents a dict, rather
than a list, and adjust BecomeUserWithoutBecomeRule so that matchplay
only runs against playbooks, not against tasks (which already have
a matchtask rule)