Handle the 'cmd' argument in command/shell
At some point, ansible grew a hidden argument to shell/command tasks,
the 'cmd' argument. History found at
https://github.com/ansible/ansible/issues/12856

Our project is making use of the cmd argument so that our shell bits do
not get jinjad, so that we can use a here doc in the command.
Unfortunately this was tripping up various tests in ansible-lint, and we
had to ignore the files, which let other errors creep in. Thus this
change adds the type of syntax we use to the test yaml files and updates
the tests to propery handle them.

Signed-off-by: Jesse Keating <jkeating@j2solutions.net>