fix skipping when using import_playbook
The `import_playbook` item doesn't have the `tasks` property, so an
exception is raised in utils.py line 669 (in `append_skipped_rules`).
Although the exception is caught and ignored the side effect is that the
rest of the method doesn't finish it's work, so the line based `noqa
XXX` statements are ignored.

For example, the following playbook should pass by default:

```yaml
- hosts: all

  tasks:
    - name: should be shell  # noqa 305 301
      shell: echo lol

- import_playbook: imported_playbook.yml
'''

Instead, the following two errors are raised:

```
ansible/playbooks/test3.yml:4: [E301] Commands should not change things if nothing needs doing
ansible/playbooks/test3.yml:4: [E305] Use shell only when shell functionality is required
```

To fix this, I added the empty list as a default value in the calls to
`.get('tasks')`.

Signed-off-by: Konstantinos Koukopoulos <koukopoulos@gmail.com>