Fixed pre-commit hooks
Assures that when used as a pre-commit hook, ansible-lint is not
given each file as argument. This is needed because ansible-lint
needs to detect the code-base layout in order to figure-out which
YAML files are playbooks, tasks or vars files.

Previously the hook was passing all files that where matching its
filters and ansible-lint would assume they were playbooks. This made it
give false positives.

This change also adds the --force-color because without it ansible-lint
would fail to detect the TTY and avoid using colors.

Fixes: #611
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>