Allow returning line number in matchplay
Currently it's not possible to set the line number ansible-lint should
show when using matchplay.
This can be a bit confusing since the line number could be very off in
case of long blocks where the line number would point to the beginning
of the block instead of the task that is faulty.

Update matchyaml to check if linenumber is also returned and use that if
it is.