Enable stub/fake modules
Provides a way to define a list of ansible modules that would be
mocked by the linter in order to allow Ansible to load the linted
files.

This should fix a very common problem which hindered linting on
various repositories, where the modules where hard to install inside
the linting jobs.

This changes also auto-adds `plugins/modules` if the folder is found,
as this is the official location for putting modules.

Fixes: #778