E208: Test with a state that would otherwise trigger the rule
The default state for the file module is state=file. This never causes a
file to be created, so the behaviour with this state (when mode is not
specified) is to preserve existing permissions. Thus in the upcoming
release of Ansible Lint it does not trigger this warning.

On the other hand state=touch will create a file if needed, thus
(correctly) triggers the warning if mode is not specified.