Fix rotten github actions
Disable broken caching on github-actions: as Github finally disabled
set-env due to security issues we are forced to disable our caching
implementation until we find a solution that works.

Switches to native py39 testing as we no longer need deadsnakes,
which are not dead for good due to the same security fix.

Related: https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/