Allow printing of double curly braces
To print double curly braces we need to write them as a jinja string in
the format {{ '{{' }}. This should not be matched by 207/Nested Jinja
Pattern as it's a valid use case.

To keep the regex as simple as possible for now we only allow double
curly braces at the beginning of the jinja string. The above example is
valid, as is {{ '{{ foo' }} but {{ 'foo {{' }} is not.