Make warn_list work as intended
If there is no config file and -w is NOT specified on the commandline
use the default value ['experimental'] for warn_list.

If there is a config file and -w is NOT specified on the commandline
use the value from the config file and discard the default.

If there is no config file and -w is specified on the commandline use
the value from the commandline and discard the default.

If there is a config file and -w is specified on the commandline merge
both values and discard the default.

Additionally document warn_list in README.rst.

Fixes #1081
Fixes #1102