Allow mocking of ansible modules
Allow user to define a list of module which will be mocked so
ansible syntax check would pass without complaining about them.

Using mocking is discouraged because it does prevent us from
identifying incorrect arguments being passed to these modules and
should be used only when installing the required modules is not
possible or too inconvenient.