Avoid looking for config file outside current git repo
This fixes a bug where running the linter inside a project that did
not had a config file could end-up picking a config file from another
project (was happening only when projects were nested).

From now on, we stop looking at parent directories when we encounter
the .git folder, which can be seen as project boundary.