Undeprecate numeric rule tags
The reason behind the undeprecation is fairly simple: right now, it is
impossible to create a playbook that uses tags that will not trigger
any warnings on ansible-lint 4 or 5.

If we use numeric tags, ansible-lint >= 5 will complain. If we use
more descriptive tags, introduced in ansible-lint 5.0.0, ansible-lint
< 4 will complain about the errors we want to ignore.

This is especially important for collection authors that publish their
work on Ansible Galaxy. On each import, Ansible Galaxy runs
ansible-lint 4 over all of the roles. Those developers are, if they
want to have a clean import, forced to use an older version of ansible
lint, or have ansible-lint >= 5 complain about numeric tags.