Allow user to specify namespace for standalone roles (#1451)
This change allows user to specify the desired namespace under which
the role is expected to be published. This allows us to better test
the role as it was installed, including performing syntax check on
test playbooks.

Previous implementation was forcing user to convert its author field
to a namespace compatible string. We no longer need this as
galaxy-importer does not have any problem when it encounters unknown
fields. Namespace is also used in collections, so it would be easier
for to remember for users.