[RFC] src/ansiblelint/_prerun.py: Add extra link to current directory in cache
ansible-lint is now checking playbook with --syntax-check, so it needs
to have the current role in cache, otherwise it will fail with error like:

internal-error: the role 'foo' was not found ...

There has been a previous attempt to fix this by creating a link to
current role base directory and the link is named according to galaxy
information. This name has several issues:
- it's specific to galaxy. People not using galaxy doesn't really need
  to respect the galaxy usage. For instance, if they follow
  https://galaxy.ansible.com/docs/contributing/creating_role.html#role-metadata
  they will put 'John Doe' in 'author' so the link will be useless
- it's uselessly breaking something which has been working before. The converge.yml
  molecule playbook is often containing only the role name (without any kind of
  namespace) and everything was fine. With the link name, people
  will have to edit every role having the issue and possibly their playbooks/
  requirements/... files too.

As a solution, this patch is also adding a link to the current role
directory name after the role name found in meta/main.yml or the
current directory name.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>