src/ansiblelint/prerun.py: Try to handle author name in metadata author
If the author field in meta/main.yml looks like "John Doe" (at least 2 words
separated by a space), act as if there was no namespace.
It's not perfect but it's not possible to match author<->galaxy account
and there will still be a warning/error issued about it, since
the fqrn won't match namespace.role_name.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>