Add rule for checking no_log is set when passwords are used (#1558)
* Add rule for checking no_log is set

It's pretty important to ensure, that password won't be logged
and thus exposed in the output. So we add linter rule that will
check that all tasks, that have "password" in argument are not logged.

Signed-Off-By: Dmitriy Rabotyagov <noonedeadpunk@ya.ru>

* Fix linters checks

Since we already have old-style unittest, leaving it as is till mass
migration to pytest. And ingoring raised by this choice warnings.

* Update NoLogPasswordsRule.py

Co-authored-by: Sorin Sbarnea <ssbarnea@redhat.com>
Co-authored-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>