Fail NoLogPassword only when loops are used
Fair note has been made in #1589 regarding modules already marking
password fields with `no_log`.
However, when loop is used, passwords are still being exposed in case
of task failure or running in debug.

So we change logic of the check and fail it only when task has some
iteration in it.