Add require_collection method
Adds utility method that detects if a collection is installed or if
it outdated and exits. This functionality is not directly used by
the linter yet but putting this code near similar prerun method makes
it easier to reuse in other related projects.

Related: https://github.com/ansible-community/molecule-podman/pull/38