Read commit description.
Because of https://github.com/ansible-collections/community.general/commit/518ace25621ee7ed08debd4d87509b0b2be33f2c, I cannot comment on this PR, so here it goes.
I will not continue working on this for obvious reasons, but feel free to use this PR if you want and close this.