E206: Increase testing accuracy
VariableHasSpacesRule testing was limited to ensuring that the number of
error was matching an expected number. This could have potential
limitation where for example a commit prevents an real linting error
from being raised and in the same time raises an invalid one.

This commit aims at improving this rule's testing by ensuring that
errors are raised only on a controlled set of lines.

Related to: #1671