Drop transformer._final_yaml_transform
There are way too many edge cases. I just ran ansible-lint --fmt-yaml on a
few thousand yaml files... And this function did not do very well. I ran
into a variety of edge cases with comments and one case where a yaml block
(|) had a line that started with #, so we got a false positive on comment
handling.

So, this approach doesn't work well enough for widespread use.