Remove support for py36 and py37
As part of reduce testing complexity and getting ansible-lint in sync
with  ansible-core, we drop support for older python versions.

This change marks the start of `6.x` and once is merged we will
create a `stable/5.x` for the last comment before it. That branch
will be used to backport changes needed for the older versions, at
least until we can make a full release of 6.x.

One additional reason for making this move is a recent explosion
of dropping support for py36 and sometimes even py37 on other
dependencies, such pytest, coverage, pip-compile, tomli.