Update tests to demonstrate issue #1507.
Note that this patch is designed to operate both before and after
the fix for issue #1507.  The assertations similar to the pre-patched
version must be the ones which fail against the pre-patched code.
If these assertations pass but others fail then the problem is likely
that the fix for issue #1507 has not been applied.