Disable _mutate_full_line_comments
It is not reliable and behaves rather oddly with many comments.
It does, however, mess up the fmt-3.yml test fixture.
I'm committing that without running prettier on it to make sure git history
demonstrates how disabling this can mess up a different set of comments.

I'll run prettier on it in a follow-up commit.