`--write`: Avoid line-wrapping YAML flow objects
When flow-objects (JSON-like YAML maps or lists) get dumped and their line
length is too long, they can get wrapped in some very odd places.

So, we calculate how long we expect the line-length might be for
flow-style maps and lists.
If it is short enough to fit on one line, we leave it in flow-style.
If the line will (probably) be wrapped, we make that
map/list use block-style instead of flow-style.