Fix "Raise name[play] for plays missing a name"
When running ansible-lint on a playbook like:

```
---
- ansible.builtin.import_playbook: "test.yml"
```

ansible-lint is now "complaining":

```
WARNING  Ignored exception from NameRule.<bound method AnsibleLintRule.matchyaml of name: Rule for checking task and play names.>: 'name'
```

The test ``in src/ansiblelint/rules/name.py:matchplay()`` is trying to ignore
the import_playbook case but with current test leads to the line:

```
data["name"], lintable=file, linenumber=data[LINE_NUMBER_KEY]
```

and in this example, the key "name" doesn't exist, leading to the warning.

The proposed solution is to suppress the special casing.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>