api: enable matchtask() to return list[MatchError] (#2445)
Extends the API for matchtask() to allow implementations to also
return list[MatchError), in addition to already supported results.

This change extends the API but does not break existing users. In
the future we might want to discourage/deprecate returning the other
type of results (bool, str or MatchError).

This change is needed in order to enable other rules to return
multiple errors within a single task file, like enabling `name` rule
to also report `name[templating]` errors.