rpm: Use %pytest and %pyproject_check_import
Use the %pytest macro to run tests and condense its invocation. It
already sets $PYTHONPATH to the destination site-packages directories.

%pyproject_check_import preforms a basic "smoke test" by importing all
public modules. It's not a replacement for unit tests, but it's better
than nothing when the tests can't run in a network-less mock chroot. The
modules that import pytest are excluded.