src/ansiblelint/file_utils.py: Fix playbook/rulebook detection
Don't try detecting the file type unless file already loaded
in Lintable._data, otherwise the line numbers for inline
rule skipping will be set to the line where's it's used since
self.kind is 'yaml'. After _guess_kind() call, the kind will
be set to playbook and the error line will be the one of the
task, which is possibly different that the one with the 'noqa:'
comment.

Fixes: #2977
Signed-off-by: Arnaud Patard <apatard@hupstream.com>