Relax ruamel-yaml requirements
As requiring the latest ruamel-yaml at runtime can impose downstream
packaging issues, we will only require a compatible version of the
library.

For testing we will only test with a fresh known version of
ruamel.yaml because we know that this impacts YAML dumping behavior.

This introduces a runtime warning when --write option is used with
older versions of ruamel.

The output of --version is also modified to include information about
versions of these two libraries.