Avoid using get_app() from inside the rule
This change does add the app object to the rule collection in order
to make it available to the rules. Also it makes the field_checks
lazy loading, so it would only be initialized much later, after
application was fully initialized.

This fixed a bug observed while running tests in parallel, where
app was initialized multiple times causing conflicts with collection
installation.