chore: allow `git rev-parse` to be used
If a user wants to call `git rev-parse`, allow this without
raising a `command-instead-of-module` error.

One use would be to do: `git rev-parse HEAD` which will get the SHA
value of the current HEAD of the git repository.

The `ansible.builtin.git` module does not have support for doing this.