Improve profile information on summary line
There are currently two problems with the message in the summary line:

If the required and passed profiles are the same, we output something
like "Profile 'production' was required, but only 'production' profile
passed." This isn't very clear because it looks like the profile was
required, but we did not manage to pass it.

Furthermore, profiles have no relation to each other. We have no way of
knowing which profile is superior to the other, so we cannot compare
them in any way.

To improve the clarity of the summary line, we remove the word 'only' in
case we do not pass the required profile, and if the required and passed
profiles are the same, we only state that the profile was passed.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>