Use checkout action in install docs
Prior to version 6.20.2, the ansible-lint action automatically included
the checkout action. It was removed in 55b8b66.

The ansible-lint action depends on the presence of a `.git` directory,
into which it downloads the `requirements-lock.txt` file. If the
directory isn't present, the action fails.

This commit updates the example YAML to work around this problem.