Accommodate specified inventory files
The ansible_cfg_inventory conditional in the _get_inventory_files method
is used because otherwise, when not passing in an inventory file via
CLI, inventory_files would set to the /etc/ansible/hosts (the default
value for the DEFAULT_HOST_LIST config). In most cases, I'd presume that
wouldn't be desired.