Do inventory parsing non-multithreaded (#4446)
This was done because the AttributeError seen in the GitHub issue
mentioned comes from ansible's codebase. ansible's API isn't
thread-safe; with how this was occurring being described here,
https://github.com/ansible/ansible/issues/13822.

run_ansible_lint is run multiple times within the
test_with_inventory_concurrent_syntax_checks test because there is some
chance that the AttributeError is not thrown on one run.