Fix inline-env-var when expand_argument_vars is used
When the expand_argument_vars parameter is used on the
ansible.builtin.command module, ansible-lint generates a false positive
with the inline-env-var rule. The expand_argument_vars parameter was
added to the ansible.builtin.command module in ansible-core 2.16 and it
appears the inline-env-var lint rule was not updated. This commit
updates the inline-env-var rule to support the expand_argument_vars
parameter.