ansible-lint-ignore: add skip option
Currently file/rules listed in ansible-lint-ignore produce a warning
in the output when matched.  This is by design, as discussed in [1],
the argument being that the error remains an error, and should be
surfaced for visibility.

The "skip_list" and "noqa" options are available to silence warnings.
There are two specific problems with this; the first acts globally and
the second can not always be applied (for example; for role-names
checks there's nowhere to put the "noqa" flag).

To use the role-name check as an example, it may be that a project
wishes to incorporate this rule moving forward, but has a selection of
existing roles that do not obey the naming rules which can not be
modified for external-API reasons.  There is no way to skip the
current failures but flag any new occurrences as issues.  Currently
the warnings are always present in reported results such as sarif
output that gets visualised by tools like GitLab.  New contributors
see a range of warnings for things the project is aware of but unable
to practically fix.

This change proposes add an optional "skip" flag to the
ansible-lint-ignore values to give no warning for the matched rule.
This is an intermediate step between the never-warn-but-global
skip_list and the always-warn ansible-lint-ignore entries, but leaves
the default behaviour to surface the warning as is currently done.

I have implemented it with the idea that rules are strictly typed by
converting users to an IgnoreRule named tuple.  This uses typed flags
with the flexibility to have other qualifiers in the future, although
the only one supported for now is "skip".

Test cases and documentation is updated.

[1] https://github.com/ansible/ansible-lint/issues/3068

Signed-off-by: Ian Wienand <iwienand@redhat.com>

# Conflicts:
#	tox.ini