Require ruamel 0.18.11 (#4615)
A change in behaviour in ruamel 0.18.11 has added an extra quote to
the end of the string match by strings.transformed.  According to
upstream this is a correct behaviour, the prior releases being
incorrect (https://sourceforge.net/p/ruamel-yaml/tickets/545/).
Consequently bump the lower-bound in requirements.in,
requirements-test.in and lower-constraints.txt and fix the match.

Everything else is the result of "tox -e deps"

Closes: https://github.com/ansible/ansible-lint/issues/4615
Signed-off-by: Ian Wienand <iwienand@redhat.com>